2012-2013 Changes
I've been iterating on the G35Arduino library since 2012 and decided to
use GitHub.  I've heavily modified the library to allow programs to be
run-time configurable.  I use ColorNodes (designed my Paul Martis at
digitalmisery.com).  The ColorNode_Controller coordinates with 30+
ColorNode_Nodes which each control individual light strings.

+++ Worm.cpp
@@ -28,7 +28,7 @@ void Worm::Do(G35& g35) {
     g35.set_color(tail_, G35::MAX_INTENSITY, COLOR_BLACK);
     tail_ += tail_dir_;
   }
-  int8_t length = abs(head_ - tail_);
+  int8_t length = abs((int)(head_ - tail_));
   if (length < UNIT) {
     is_stretching_ = true;
   }
